change the entrypoint to the argparse, so we can pass arguments to the executable.
