+++ src/tmeanalyzer/__init__.py
@@ -1,3 +1,2 @@
 from .version import __version__
-from .start import start
-from .TME_analyzer import check_consent
+from .TME_Analyzer import start
